refactor(reposicao): split AdminReposicaoAplicacao (974→656) extraindo leaf components
Quebra do componente grande seguindo o padrão de PromocaoDetail/FinanceiroDashboard.
Move para src/components/reposicao/aplicacao/:
- types.ts (FilaItem, FilaCounterRow, GerarFilaResult, RegistrarSubstResult,
  SkuParametroOpcao, EMPRESA)
- DeltaArrow.tsx (deltaPct module-local + DeltaArrow)
- SubstituicaoPendenteCard.tsx (card auto-contido com queries/mutations próprias)
- SubstituicaoModal.tsx (modal auto-contido com state/queries/mutations)

Movimentação 1:1 verbatim, comportamento preservado. Imports mortos removidos da
página (Textarea, Label, RadioGroup, bloco Dialog, ArrowRight). Página 974 → 656.
tsc limpo, 452/452 testes.

Follow-up (deferido): extrair os 4 TabsContent (pronto/inativo/substituicao/aplicado)
— maior risco de prop-drilling.
