+++ src/pages/AdminReposicaoAplicacao.tsx
@@ -5,7 +5,6 @@ import { toast } from "sonner";
 import { format, formatDistanceToNow } from "date-fns";
 import { ptBR } from "date-fns/locale";
 import {
-  ArrowRight,
   RefreshCw,
   PlayCircle,
   AlertTriangle,
@@ -27,17 +26,6 @@ import {
   SelectValue,
 } from "@/components/ui/select";
 import { Input } from "@/components/ui/input";
-import { Textarea } from "@/components/ui/textarea";
-import { Label } from "@/components/ui/label";
-import { RadioGroup, RadioGroupItem } from "@/components/ui/radio-group";
-import {
-  Dialog,
-  DialogContent,
-  DialogHeader,
-  DialogTitle,
-  DialogFooter,
-  DialogDescription,
-} from "@/components/ui/dialog";
 import {
   AlertDialog,
   AlertDialogAction,
@@ -56,82 +44,15 @@ import {
   TableHeader,
   TableRow,
 } from "@/components/ui/table";
-
-const EMPRESA = "OBEN";
-
-// Tipos locais — tabelas custom (fila_aplicacao_omie, sku_substituicao) e RPCs
-// (gerar_fila_aplicacao_omie, registrar_substituicao_sku) ainda não estão no
-// Database type gerado. Quando entrarem, podem ser substituídos por Tables<>.
-type FilaItem = {
-  id: number;
-  empresa: string;
-  sku_codigo_omie: string;
-  sku_descricao: string | null;
-  estoque_minimo_novo: number | null;
-  ponto_pedido_novo: number | null;
-  estoque_maximo_novo: number | null;
-  estoque_minimo_omie_atual: number | null;
-  ponto_pedido_omie_atual: number | null;
-  estoque_maximo_omie_atual: number | null;
-  status_validacao: string;
-  mensagem_bloqueio: string | null;
-  delta_max_perc: number | null;
-  aplicado_em: string | null;
-  resposta_omie: Record<string, unknown> | null;
-  erro_omie: string | null;
-  criado_em: string;
-};
-
-interface FilaCounterRow {
-  status_validacao: string;
-  aplicado_em: string | null;
-}
-
-interface GerarFilaResult {
-  prontos?: number;
-  bloqueados_inativos?: number;
-  bloqueados_substituicao?: number;
-}
-
-interface RegistrarSubstResult {
-  error?: string;
-}
-
-interface SkuParametroOpcao {
-  sku_codigo_omie: number;
-  sku_descricao: string | null;
-}
-
-function deltaPct(novo: number | null, atual: number | null): number | null {
-  if (novo == null) return null;
-  if (atual == null || atual === 0) return null;
-  return ((novo - atual) / atual) * 100;
-}
-
-function DeltaArrow({ novo, atual }: { novo: number | null; atual: number | null }) {
-  const pct = deltaPct(novo, atual);
-  return (
-    <div className="flex items-center gap-1 text-xs whitespace-nowrap">
-      <span className="text-muted-foreground">{atual ?? "—"}</span>
-      <ArrowRight className="h-3 w-3 text-muted-foreground" />
-      <span className="font-medium">{novo ?? "—"}</span>
-      {pct != null && (
-        <span
-          className={
-            Math.abs(pct) > 25
-              ? "text-destructive ml-1"
-              : Math.abs(pct) > 10
-              ? "text-warning ml-1"
-              : "text-muted-foreground ml-1"
-          }
-        >
-          ({pct > 0 ? "+" : ""}
-          {pct.toFixed(0)}%)
-        </span>
-      )}
-    </div>
-  );
-}
+import {
+  EMPRESA,
+  type FilaItem,
+  type FilaCounterRow,
+  type GerarFilaResult,
+} from "@/components/reposicao/aplicacao/types";
+import { DeltaArrow } from "@/components/reposicao/aplicacao/DeltaArrow";
+import { SubstituicaoPendenteCard } from "@/components/reposicao/aplicacao/SubstituicaoPendenteCard";
+import { SubstituicaoModal } from "@/components/reposicao/aplicacao/SubstituicaoModal";
 
 export default function AdminReposicaoAplicacao() {
   const qc = useQueryClient();
@@ -731,244 +652,5 @@ export default function AdminReposicaoAplicacao() {
         </AlertDialogContent>
       </AlertDialog>
     </div>
-  );
-}
-
-/* ============ Componentes auxiliares ============ */
-
-function SubstituicaoPendenteCard({
-  item,
-  onChange,
-}: {
-  item: FilaItem;
-  onChange: () => void;
-}) {
-  const { data: subst } = useQuery({
-    queryKey: ["sku-substituicao", item.empresa, item.sku_codigo_omie],
-    queryFn: async () => {
-      const { data } = await supabase
-        .from("sku_substituicao")
-        .select("*")
-        .eq("empresa", item.empresa)
-        .eq("sku_codigo_antigo", item.sku_codigo_omie)
-        .eq("status", "pendente")
-        .maybeSingle();
-      return data;
-    },
-  });
-
-  const cancelarSubst = useMutation({
-    mutationFn: async () => {
-      const { error } = await supabase
-        .from("sku_substituicao")
-        .update({ status: "cancelada" } as never)
-        .eq("id", subst!.id);
-      if (error) throw error;
-    },
-    onSuccess: () => {
-      toast.success("Substituição cancelada");
-      onChange();
-    },
-  });
-
-  const aplicarSubst = useMutation({
-    mutationFn: async () => {
-      const { error } = await supabase
-        .from("sku_substituicao")
-        .update({ status: "aplicada", aplicado_em: new Date().toISOString() } as never)
-        .eq("id", subst!.id);
-      if (error) throw error;
-    },
-    onSuccess: () => {
-      toast.success("Substituição aplicada. Regere a fila para revalidar.");
-      onChange();
-    },
-  });
-
-  return (
-    <Card className="border-warning/40">
-      <CardContent className="py-4 space-y-3">
-        <div className="flex items-start justify-between gap-3">
-          <div>
-            <div className="font-mono text-sm">{item.sku_codigo_omie} (antigo)</div>
-            <div className="font-medium">{item.sku_descricao}</div>
-            {subst && (
-              <div className="mt-2 text-xs">
-                <div>
-                  <span className="text-muted-foreground">SKU novo: </span>
-                  <span className="font-mono">{subst.sku_codigo_novo}</span>
-                </div>
-                <div>
-                  <span className="text-muted-foreground">Ação: </span>
-                  <Badge variant="outline">{subst.acao_parametros}</Badge>
-                </div>
-                <div className="text-muted-foreground mt-1">
-                  Motivo: {subst.motivo ?? "—"}
-                </div>
-              </div>
-            )}
-          </div>
-          <Badge className="bg-warning/20 text-warning-foreground">Substituição pendente</Badge>
-        </div>
-        <div className="flex gap-2">
-          <Button size="sm" onClick={() => aplicarSubst.mutate()} disabled={!subst}>
-            Aplicar substituição
-          </Button>
-          <Button
-            size="sm"
-            variant="outline"
-            onClick={() => cancelarSubst.mutate()}
-            disabled={!subst}
-          >
-            Cancelar substituição
-          </Button>
-        </div>
-      </CardContent>
-    </Card>
-  );
-}
-
-function SubstituicaoModal({
-  item,
-  onClose,
-  onDone,
-}: {
-  item: FilaItem;
-  onClose: () => void;
-  onDone: () => void;
-}) {
-  const [skuNovo, setSkuNovo] = useState("");
-  const [busca, setBusca] = useState("");
-  const [acao, setAcao] = useState("transferir");
-  const [motivo, setMotivo] = useState("");
-
-  const { data: opcoes } = useQuery({
-    queryKey: ["sku-busca", busca],
-    queryFn: async () => {
-      if (!busca || busca.length < 2) return [];
-      const { data } = await supabase
-        .from("sku_parametros")
-        .select("sku_codigo_omie, sku_descricao")
-        .eq("empresa", EMPRESA)
-        .or(`sku_codigo_omie.eq.${Number(busca) || 0},sku_descricao.ilike.%${busca}%`)
-        .limit(20);
-      return data ?? [];
-    },
-    enabled: busca.length >= 2,
-  });
-
-  const registrar = useMutation({
-    mutationFn: async () => {
-      if (!skuNovo) throw new Error("Selecione o SKU novo");
-      if (!motivo.trim()) throw new Error("Motivo é obrigatório");
-      const { data: { user } } = await supabase.auth.getUser();
-      const { data, error } = await supabase.rpc("registrar_substituicao_sku" as never, {
-        p_empresa: item.empresa,
-        p_codigo_antigo: item.sku_codigo_omie,
-        p_codigo_novo: skuNovo,
-        p_acao_parametros: acao,
-        p_motivo: motivo,
-        p_usuario: user?.email ?? "sistema",
-      } as never);
-      if (error) throw error;
-      const result = data as unknown as RegistrarSubstResult | null;
-      if (result?.error) throw new Error(result.error);
-      return data;
-    },
-    onSuccess: () => {
-      toast.success("Substituição registrada");
-      onDone();
-    },
-    onError: (e: Error) => toast.error(e.message),
-  });
-
-  return (
-    <Dialog open onOpenChange={(o) => !o && onClose()}>
-      <DialogContent className="max-w-lg">
-        <DialogHeader>
-          <DialogTitle>Registrar substituição</DialogTitle>
-          <DialogDescription>
-            SKU antigo: <span className="font-mono">{item.sku_codigo_omie}</span> —{" "}
-            {item.sku_descricao}
-          </DialogDescription>
-        </DialogHeader>
-        <div className="space-y-4">
-          <div>
-            <Label>SKU novo</Label>
-            <Input
-              placeholder="Buscar por código ou descrição"
-              value={busca}
-              onChange={(e) => setBusca(e.target.value)}
-            />
-            {opcoes && opcoes.length > 0 && (
-              <div className="mt-2 border rounded max-h-44 overflow-auto text-sm">
-                {(opcoes as SkuParametroOpcao[]).map((o) => (
-                  <button
-                    key={o.sku_codigo_omie}
-                    type="button"
-                    onClick={() => {
-                      setSkuNovo(String(o.sku_codigo_omie));
-                      setBusca(`${o.sku_codigo_omie} — ${o.sku_descricao}`);
-                    }}
-                    className={`block w-full text-left px-3 py-1.5 hover:bg-muted ${
-                      skuNovo === String(o.sku_codigo_omie) ? "bg-muted" : ""
-                    }`}
-                  >
-                    <span className="font-mono">{o.sku_codigo_omie}</span> — {o.sku_descricao}
-                  </button>
-                ))}
-              </div>
-            )}
-            {skuNovo && (
-              <p className="text-xs text-muted-foreground mt-1">Selecionado: {skuNovo}</p>
-            )}
-          </div>
-
-          <div>
-            <Label>Ação sobre parâmetros</Label>
-            <RadioGroup value={acao} onValueChange={setAcao} className="mt-2">
-              <div className="flex items-start gap-2">
-                <RadioGroupItem value="transferir" id="r1" className="mt-1" />
-                <Label htmlFor="r1" className="font-normal">
-                  <span className="font-medium">Transferir</span> — copia parâmetros do antigo para
-                  o novo e aprova o novo.
-                </Label>
-              </div>
-              <div className="flex items-start gap-2">
-                <RadioGroupItem value="recalcular_do_zero" id="r2" className="mt-1" />
-                <Label htmlFor="r2" className="font-normal">
-                  <span className="font-medium">Recalcular do zero</span> — sistema calcula a
-                  partir do histórico do novo SKU.
-                </Label>
-              </div>
-              <div className="flex items-start gap-2">
-                <RadioGroupItem value="manter_ambos" id="r3" className="mt-1" />
-                <Label htmlFor="r3" className="font-normal">
-                  <span className="font-medium">Manter ambos</span> — registra mas não desativa o
-                  antigo.
-                </Label>
-              </div>
-            </RadioGroup>
-          </div>
-
-          <div>
-            <Label>Motivo *</Label>
-            <Textarea
-              value={motivo}
-              onChange={(e) => setMotivo(e.target.value)}
-              placeholder="Ex.: descontinuação do fornecedor, troca de embalagem, etc."
-            />
-          </div>
-        </div>
-        <DialogFooter>
-          <Button variant="outline" onClick={onClose}>
-            Cancelar
-          </Button>
-          <Button onClick={() => registrar.mutate()} disabled={registrar.isPending}>
-            Confirmar
-          </Button>
-        </DialogFooter>
-      </DialogContent>
-    </Dialog>
   );
 }
